optimise remove_containing_nodes by learning lower bound
The MAXRECTS-BSSF algorithm has three major steps:

- 'scan_bins' looks at all free rectangles to find the best
  short side fit.
- 'split_intersecting_nodes' performs intersection tests
  between  all free rectangles and the chosen rectangle,
  spliting any that intersect. This has O(n) complexity.
- 'remove_containing_nodes' performs contains tests on all
  ordered pairs of rectangles, removing any contained
  rectangles. This pass has O(n^2) complexity.

It can be observed that the containing rectangle tests in
'remove_containing_nodes' will produce the same results for
the same set of rectangles, on subsequent iterations, as they
have no changed, so we can record the lower bound from where
it is known that all rectangles do not contain any other
rectangle, and use that lower bound as a starting point to
accelerate this pass on subsequent iterations. In addition to
recording this lower bound in the first pass, we also need to
invalidate the known lower bound when 'split_intersecting_nodes'
makes changes to the free list.